added hp text and bg music to battle scene

+++ HistoriaFamilia/Assets/Scripts/UnitType.cs
@@ -24,7 +24,7 @@ public class UnitType {
 
 	public int MaxHealth = 10;
 	public int MovementReach = 6;
-	public float BaseAttackPower = 5;
+	public int BaseAttackPower = 5;
 	public float AttackRange = 1;
  	public List<TileType.TopographicalFeature> ProhibitedToWalkOn;
 }
